% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFSestimates-package.R
\docType{package}
\name{BFSestimates-package}
\alias{BFSestimates}
\alias{BFSestimates-package}
\title{BFSestimates: Estimate totals and confidence intervals from Bundesamt für Statistik's Surveys}
\description{
Estimates population sizes and confidence intervals from "Bundesamt für Statistik (BFS)" / "Office fédéral de la statistique (OFS)" population and mobility surveys. Give it data from "Strukturerhebung" / "relevé structurel" or from "Mikrozensus für Mobilität und Verkehr" / "Microrecensement mobilité et transports", and obtain estimates of totals and confidence intervals.
}
\keyword{internal}
