% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chensus-deprecated.R, R/deprecated_wrappers.R
\name{chensus-deprecated}
\alias{chensus-deprecated}
\alias{se_estimate_total}
\alias{se_estimate_mean}
\alias{mzmv_estimate_mean}
\alias{mzmv_estimate_mean_map}
\title{Deprecated functions in chensus}
\usage{
se_estimate_total(...)

se_estimate_mean(...)

mzmv_estimate_mean(...)

mzmv_estimate_mean_map(...)
}
\description{
These functions are provided for compatibility with older versions of `chensus`.
They may be removed in a future release.
}
\keyword{internal}
