% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chensus-package.R
\docType{package}
\name{chensus-package}
\alias{chensus}
\alias{chensus-package}
\title{chensus: Estimate Totals, Means, Proportions and Confidence Intervals of the Federal Statistic Office's Surveys}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimates population totals, means, proportions and confidence intervals from the Federal Statistic Office's (FSO) structural and mobility surveys. Give it data from "Strukturerhebung" / "relevé structurel" or from "Mikrozensus für Mobilität und Verkehr" / "Microrecensement mobilité et transports", and obtain estimates of totals, means, proportions and confidence intervals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/afds-bl/chensus}
}

}
\author{
\strong{Maintainer}: Souad Guemghar \email{souad.guemghar@bl.ch}

Other contributors:
\itemize{
  \item Amt für Daten und Statistik, Basel-Landschaft [copyright holder, funder]
}

}
\keyword{internal}
