% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fso_mask_flag.R
\name{fso_flag_mask}
\alias{fso_flag_mask}
\title{Classify Estimate Reliability and Apply Confidentiality Masking}
\usage{
fso_flag_mask(data, lang = c("de", "fr", "it", "en"))
}
\arguments{
\item{data}{A data frame or tibble.}

\item{lang}{A character string for the language of the estimate reliability description, one of "de", "fr", "it", "en". 
Defaults to German if omitted.}
}
\value{
A tibble containing the original data with masked estimates when \code{occ <= 4} and one additional column:
\describe{
  \item{obs_status}{characater column classifying reliability of estimates.}
}
}
\description{
\code{fso_flag_mask} applies Swiss Federal Statistical Office (FSO) reliability rules for survey estimates,
based on the number of observations (\code{occ}). It flags low reliability estimates and masks them when sample size is too small 
(\code{occ <= 4}).
}
\details{
FSO estimate reliability criteria:
\describe{
  \item{\code{occ <= 4: }}{No estimate (confidential).}
  \item{\code{occ <= 49: }}{Estimate of low reliability.}
  \item{\code{occ > 49: }}{Reliable estimate.}
}
}
\examples{
df <- data.frame(occ = c(3, 10, 60), mean_income = c(4000, 4200, 4500))
fso_flag_mask(df)

}
