% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzmv_estimate.R
\name{mzmv_estimate_mean}
\alias{mzmv_estimate_mean}
\title{Estimate means from mobility survey}
\usage{
mzmv_estimate_mean(data, object, weight, cf = 1.14, alpha = 0.1)
}
\arguments{
\item{data}{Tibble}

\item{object}{Vector of strings, names of variables to be estimated. Variables have integer values, representing a quantity (number of cars per household) or presence/absence (possession of a car). Negative numbers represent `NA`.}

\item{weight}{Character string, name of the column containing the
weights}

\item{cf}{Double, correction factor of the confidence interval, supplied by BFS/OFS}

\item{alpha}{Double, significance level. Default 0.05 for 95\% confidence interval.}
}
\value{
Tibble (number of rows is length of \code{object}) with the following columns:
\itemize{
\item{\code{id}: }{estimated item}
\item{\code{nc}: }{number of survey responses}
\item{\code{wmean}: }{weighted mean estimate}
\item{\code{ci}: }{confidence interval estimate}
}
}
\description{
\code{mzmv_estimate_mean()} estimates the mean frequencies and confidence
intervals of BFS/OFS mobility surveys.
}
\examples{
# We can use the nhanes dataset as an example even if it only contains population data
library(dplyr)
# Estimate two means
mzmv_estimate_mean(
  c("annual_household_income", "annual_family_income"),
  data = nhanes,
  weight = "weights",
  alpha = 0.1
)
# With conditions
c("gender", "interview_lang") \%>\%
  purrr::set_names() \%>\%
  purrr::map(~ mzmv_estimate_mean(
    object = c("annual_household_income", "annual_family_income"),
    data = nhanes \%>\% group_by(.data[[.x]]), weight = "weights"
  ))

}
