% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzmv_estimate.R
\name{mzmv_estimate_prop}
\alias{mzmv_estimate_prop}
\title{Estimate proportions from mobility survey}
\usage{
mzmv_estimate_prop(data, variable, weight, cf = 1.14, alpha = 0.1)
}
\arguments{
\item{data}{Tibble}

\item{variable}{Vector of strings, names of variables to be estimated. Variables are binary with integer values:
\itemize{
\item 1: if condition is present
\item 0: if condition is absent
\item negative: if \code{NA}
}}

\item{weight}{Character string, name of the column containing the
weights}

\item{cf}{Double, correction factor of the confidence interval, supplied by FSO}

\item{alpha}{Double, significance level. Default 0.05 for 95\% confidence interval.}
}
\value{
Vector, with the following values:
\itemize{
\item \code{p}: proportion estimate
\item \code{ci}: confidence interval estimate
}
}
\description{
\code{mzmv_estimate_prop} estimates the proportions and confidence intervals of FSO mobility survey data
}
\examples{
# We can use the nhanes dataset
library(dplyr)
nhanes \%>\%
  mutate(
    married =
      case_when(
        marital_status == "Married" ~ 1,
        TRUE ~ 0
      )
  ) \%>\%
  mzmv_estimate_prop(
    variable = "married",
    weight = "weights"
  )

}
