% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes.R
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{National Health and Nutrition Examination Survey (NHANES)}
\format{
A data frame with 9307 rows and 13 variables.
}
\source{
\href{https://wwwn.cdc.gov/Nchs/Nhanes/2015-2016/DEMO_I.htm}{NHANES 2015-2016}
}
\usage{
nhanes
}
\description{
Demographic survey data from NHANES 2015 to 2016, with data on 9307
participants, including sampling weights.
}
\examples{
library(dplyr)
glimpse(nhanes)
nhanes \%>\% count(edu_level)
}
\references{
(\href{https://wwwn.cdc.gov/nchs/nhanes/}{CDC})
}
\keyword{datasets}
