% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_dummy.R
\name{se_dummy}
\alias{se_dummy}
\title{Create dummy variables}
\usage{
se_dummy(data, column, id)
}
\arguments{
\item{data}{Tibble}

\item{column}{Character string, name of variable to dummify}

\item{id}{Character string, name of unique identifier ID column}
}
\value{
Tibble composed of original tibble and newly created dummy variables
}
\description{
Create dummy variables
}
\examples{
data <- data.frame(
id = 1:5,
category = c("A", "B", "A", "C", "B")
)
se_dummy(data, "category", "id")

}
