% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_estimate_mean.R
\name{se_estimate_mean}
\alias{se_estimate_mean}
\title{Estimate means of population surveys}
\usage{
se_estimate_mean(
  data,
  variable,
  var_type,
  strata = "zone",
  weight,
  condition = NULL,
  alpha = 0.05
)
}
\arguments{
\item{data}{Tibble}

\item{variable}{Character string, name of the feature whose mean we would like
to estimate}

\item{var_type}{Charcter string, variable type, one of "cat" or "num"}

\item{strata}{Character string, name of the column containing the
strata/zones}

\item{weight}{Character string, name of the column containing the
weights}

\item{condition}{condition Vector of character strings, names of additional
stratification variables}

\item{alpha}{Double, significance level. Default 0.05 for 95\% confidence interval.}
}
\value{
Tibble, with the following columns:
 \itemize{
 \item{\code{occ}: }{true frequency in survey sample}
 \item{\code{average}: }{estimated mean}
 \item{\code{vhat}: }{estimated variance}
 \item{\code{stand_dev}: }{standard deviation}
 \item{\code{ci}: }{absolute confidence interval}
 }
}
\description{
\code{se_estimate_mean()} estimates the proportions of categorical variables
and averages of numeric variables along with the variance and confidence
intervals of the Strukturerhebung / relevé struturel of BFS/OFS/FSO.
}
\examples{
se_estimate_mean(
  data = nhanes,
  variable = "age",
  var_type = "num",
  strata = "strata",
  weight = "weights",
  condition = "gender"
)

}
