% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_estimate_total.R
\name{se_estimate_total}
\alias{se_estimate_total}
\title{Estimates from population survey}
\usage{
se_estimate_total(
  data,
  weight,
  strata = "zone",
  condition = NULL,
  alpha = 0.05
)
}
\arguments{
\item{data}{Tibble}

\item{weight}{Character string, name of the column containing the
weights}

\item{strata}{Character string, name of the column containing the
strata/zones}

\item{condition}{Vector of character strings, names of the conditions to
estimate, can be empty for total population estimate}

\item{alpha}{Double, significance level. Default 0.05 for 95\% confidence interval.}
}
\value{
Tibble, with the following columns:
 \itemize{
 \item{\code{total}: }{population estimate}
 \item{\code{vhat}: }{estimated variance}
 \item{\code{occ}: }{true frequency in survey sample}
 \item{\code{stand_dev}: }{standard deviation}
 \item{\code{ci}: }{absolute confidence interval}
 \item{\code{ci_per}: }{percent confidence interval.}
 }
}
\description{
\code{se_estimate()} estimates the frequencies, variance and confidence
intervals of BFS/OFS population surveys.
}
\examples{
# One condition
se_estimate_total(
  data = nhanes,
  weight = "weights",
  strata = "strata",
  condition = "gender"
)
# Multiple conditions
library(dplyr)
library(purrr)
map(
  c("gender", "marital_status"),
  ~ se_estimate_total(
    data = nhanes,
    weight = "weights",
    strata = "strata",
    condition = .x
  ) \%>\%
    mutate(variable = .x, .before = 1) \%>\%
    rename_with(~"value", all_of(.x))
) \%>\%
  map_dfr(~ .x \%>\% as_tibble())

}
