% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_mean.R
\name{se_mean_cat}
\alias{se_mean_cat}
\title{Estimate means of categorical variables in structural survey}
\usage{
se_mean_cat(
  data,
  variable,
  group_vars = NULL,
  condition = NULL,
  strata = "zone",
  weight,
  alpha = 0.05
)
}
\arguments{
\item{data}{Tibble}

\item{variable}{Character string, name of the categorical variable whose proportion we would like
to estimate}

\item{group_vars}{A character vector of grouping variables.}

\item{condition}{[Deprecated] Use `group_vars` instead. A character vector of grouping variables.}

\item{strata}{Character string, name of the column containing the
strata/zones}

\item{weight}{Character string, name of the column containing the
weights}

\item{alpha}{Double, significance level. Default 0.05 for 95\% confidence interval.}
}
\value{
Tibble, with the following columns:
 \itemize{
 \item \code{occ}: true frequency in survey sample
 \item \code{average}: estimated mean
 \item \code{vhat}: estimated variance
 \item \code{stand_dev}: standard deviation
 \item \code{ci}: absolute confidence interval
 }
}
\description{
\code{se_mean_cat()} estimates the proportion of categorical variables along with the variance and confidence
intervals of FSO's structural survey (Strukturerhebung / relevé structurel).
}
\examples{
se_mean_cat(
  data = nhanes,
  variable = "age",
  strata = "strata",
  weight = "weights",
  group_vars = "gender"
)

}
