% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_ogd.R
\name{se_mean_ogd}
\alias{se_mean_ogd}
\title{Estimate Means for All Combinations of Grouping Variables (OGD Format)}
\usage{
se_mean_ogd(data, variable, ..., strata, weight, alpha = 0.05)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{variable}{Variable to estimate the mean for (unquoted or programmatic).}

\item{...}{Grouping variables (unquoted or programmatic).}

\item{strata}{Stratification variable (unquoted or programmatic). Defaults to "zone" if omitted.}

\item{weight}{Sampling weights variable (unquoted or programmatic).}

\item{alpha}{Significance level for confidence intervals. Default is 0.05 (95\% CI).}
}
\value{
A tibble with survey mean estimates for all combinations of grouping variables.
  Grouping variables are converted to factors with "Total" representing the overall group.
}
\description{
\code{se_mean_ogd} estimates survey means of a continuous variable for every combination of the supplied grouping variables,
using \code{se_mean} internally and returning results in a format suitable for Open Government Data (OGD).
The output includes means for each combination of grouping variables, as well as for the overall population.
}
\examples{
# Unquoted variables
se_mean_ogd(nhanes, variable = household_size, strata = strata, weight = weights, gender)

# Programmatic use
var <- "household_size"
wt <- "weights"
vars <- "gender"
se_mean_ogd(
  nhanes,
  variable = !!rlang::sym(var),
  strata = strata,
  weight = !!rlang::sym(wt),
  !!!rlang::syms(vars)
)

}
\seealso{
\code{\link{se_total_ogd}}, \code{\link{se_prop_ogd}}, \code{\link{se_ogd_wrapper}}, \code{\link{se_mean}}
}
