% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_ogd.R
\name{se_ogd_wrapper}
\alias{se_ogd_wrapper}
\title{OGD Wrapper for Structural Survey Estimation Functions}
\usage{
se_ogd_wrapper(
  data,
  core_fun,
  ...,
  strata,
  weight,
  alpha = 0.05,
  variable = NULL,
  show_internal = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{core_fun}{The core estimation function to use, one of \code{se_mean}, \code{se_total}, \code{se_prop}.}

\item{...}{Grouping variables (unquoted or programmatic).}

\item{strata}{Stratification variable (unquoted or programmatic).}

\item{weight}{Sampling weights variable (unquoted or programmatic).}

\item{alpha}{Significance level for confidence intervals.}

\item{variable}{(Optional) Variable to estimate mean for (only needed for se_mean).}

\item{show_internal}{Show internal estimates of variance, standard deviation (and percent confidence interval for \code{se_total()}).
Hidden by default.}
}
\value{
A tibble with estimates for all combinations of grouping variables.
}
\description{
OGD Wrapper for Structural Survey Estimation Functions
}
