% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_total_comb.R
\name{se_total_comb}
\alias{se_total_comb}
\title{Calculate Estimates for All Combinations of Grouping Variables in Structural Survey}
\usage{
se_total_comb(data, ..., strata, weight, alpha = 0.05)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{...}{Optional grouping variables. Can be passed unquoted (e.g., \code{gender}, \code{birth_country}) or programmatically using \code{!!!syms(c("gender", "birth_country"))}..}

\item{strata}{Unquoted or quoted name of the strata column. Defaults to \code{zone} if omitted.}

\item{weight}{Unquoted or quoted name of the sampling weights column. For programmatic use.
with a string variable (e.g., \code{wt <- "weights"}), use \code{!!sym(wt)} in the function call.}
}
\value{
A tibble combining survey estimates for all combinations of \code{group_vars},
  with grouping variables converted to factors where "Total" means .
}
\description{
\code{se_total_comb()} computes structural survey totals for every combination of the supplied grouping variables.
It uses \code{se_total()} and \code{se_combn()} internally and returns a combined tibble with results for each grouping subset,
including totals across all groups. This formatting can be useful to publish on Open Government Data platforms.
}
\examples{
# Unquoted variables
se_total_comb(nhanes, strata = strata, weight = weights, gender, birth_country)

# Programmatic use
wt <- "weights"
strata <- "strata"
vars <- c("gender", "birth_country")

se_total_comb(
  nhanes,
  strata = "strata",
  weight = "weights",
  !!!rlang::syms(vars)
)

}
\seealso{
\code{\link[=se_total]{se_total()}}, \code{\link[=se_total_map]{se_total_map()}}
}
