% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_total_prop.R
\name{se_total_prop}
\alias{se_total_prop}
\title{Format Table with Totals and Proportions for a Categorical Variable}
\usage{
se_total_prop(data, ..., strata, weight, alpha = 0.05)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{...}{Optional grouping variables (unquoted).}

\item{strata}{The name of the strata variable (default is "zone").}

\item{weight}{The name of the weight variable.}

\item{alpha}{Significance level for confidence intervals. Default is 0.05.}
}
\value{
A tibble with proportions and totals joined, masked as necessary.
}
\description{
This function estimates proportions for a categorical variable grouped by optional variables, computes totals,
joins both estimates, removes unnecessary columns, and flags masked values.
}
\examples{
se_total_prop(
  data = nhanes,
  interview_lang,
  gender,
  birth_country,
  strata = strata,
  weight = weights
)

}
