% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vz_summarise.R
\name{vz_summarise}
\alias{vz_summarise}
\title{Summarise population survey}
\usage{
vz_summarise(data, weight, strata, mh_col = "mh", Nh_col = "Nh")
}
\arguments{
\item{data}{Tibble}

\item{weight}{Character string, name of the column containing the
weights}

\item{strata}{Vector of character strings, names of the columns containing the
strata}

\item{mh_col}{Character string, desired column name of number of true occurrences per stratum}

\item{Nh_col}{Character string, desired column name of estimated population per stratum}
}
\value{
Tibble \code{data} augmented with columns \code{mh_col} and \code{Nhcol}
}
\description{
\code{vz_summarise()} stratifies the number of true occurrences and
estimated populations in population surveys provided by the the Bundesamt
für Statistik / Office Fédéral de la Statistique.
}
\examples{
# One strata variable
library(dplyr)
vz_summarise(nhanes, weight = "weights", strata = "strata") \%>\%
  glimpse()
# Two strata variables
vz_summarise(nhanes,
  weight = "weights", strata = c("strata", "gender"),
  mh_col = "mhc", Nh_col = "Nhc"
) \%>\%
  select(weights, strata, gender, mhc, Nhc) \%>\%
  group_by(strata, gender) \%>\%
  slice_sample(n = 2) \%>\%
  head(10)

}
