% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFSestimates-package.R
\docType{package}
\name{BFSestimates-package}
\alias{BFSestimates}
\alias{BFSestimates-package}
\title{BFSestimates: Estimate Populations from Bundesamt für Statistik Surveys}
\description{
Estimates population sizes and confidence intervals from Bundesamt für Statistik (BFS) / Office fédéral de la statistique (OFS) population and mobility surveys. Give it data from Volkszählung / Recensement de la population or from Mikrozensus für Mobilität und Verkehr / Microrecensement mobilité et transports, and obtain population estimates along with confidence intervals.
}
\keyword{internal}
