% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFSestimates-package.R
\docType{package}
\name{BFSestimates-package}
\alias{BFSestimates}
\alias{BFSestimates-package}
\title{BFSestimates: Estimate Totals, Averages and Confidence Intervals of the Federal Statistic Office's Surveys}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimates population sizes, averages and confidence intervals from the Federal Statistic Office's (FSO) population and mobility surveys. Give it data from "Strukturerhebung" / "relevé structurel" or from "Mikrozensus für Mobilität und Verkehr" / "Microrecensement mobilité et transports", and obtain estimates of totals and confidence intervals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/afds-bl/bfsestimates}
}

}
\author{
\strong{Maintainer}: Souad Guemghar \email{souad.guemghar@bl.ch}

Other contributors:
\itemize{
  \item Amt für Daten und Statistik, Basel-Landschaft [copyright holder, funder]
}

}
\keyword{internal}
