% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzmv_estimate.R
\name{mzmv_mean}
\alias{mzmv_mean}
\title{Estimate Means of Mobility Survey}
\usage{
mzmv_mean(data, ..., weight, cf = 1.14, alpha = 0.1)
}
\arguments{
\item{data}{A tibble or data frame.}

\item{...}{Unquoted names of variables to be estimated. Variables have integer values, representing a quantity (number of cars per household) or presence/absence (possession of a car). Negative numbers represent `NA`.}

\item{weight}{Unquoted name of the column containing the sample
weights.}

\item{cf}{Double, correction factor of the confidence interval, supplied by FSO. Default is 1.14.}

\item{alpha}{Double, significance level. Default 0.1 for 90\% confidence interval.}
}
\value{
Tibble (number of rows is length of \code{variable}) with the following columns:
\itemize{
\item \code{id}: estimated item
\item \code{nc}: number of survey responses
\item \code{wmean}: weighted mean estimate
\item \code{ci}: confidence interval estimate
}
}
\description{
\code{mzmv_mean()} estimates the mean frequencies and confidence
intervals of FSO mobility surveys.
}
\examples{
library(dplyr)
library(purrr)
# Estimate two means
mzmv_mean(
data = nhanes,
annual_household_income, annual_family_income,
weight = weights
)
}
\seealso{
See \code{\link{mzmv_mean_map}} for estimates on a set of conditions.
}
