% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzmv_estimate.R
\name{mzmv_mean_map}
\alias{mzmv_mean_map}
\title{Estimate means of mobility survey with conditions}
\usage{
mzmv_mean_map(
  data,
  variable,
  condition = NULL,
  group_vars = NULL,
  weight,
  cf = 1.14,
  alpha = 0.1
)
}
\arguments{
\item{data}{Tibble}

\item{variable}{Vector of strings, names of variables to be estimated. Variables have integer values, representing a quantity (number of cars per household) or presence/absence (possession of a car). Negative numbers represent `NA`.}

\item{group_vars}{Vector of character strings, names of additional
stratification variables}

\item{weight}{Character string, name of the column containing the
weights}

\item{cf}{Double, correction factor of the confidence interval, supplied by FSO}

\item{alpha}{Double, significance level. Default 0.1 for 90\% confidence interval.}
}
\value{
Tibble (number of rows is length of \code{variable}) with the following columns:
\itemize{
\item \code{id}: estimated item
\item \code{nc}: number of survey responses
\item \code{wmean}: weighted mean estimate
\item \code{ci}: confidence interval estimate
}
}
\description{
\code{mzmv_mean_map()} estimates the mean frequencies and confidence
intervals of FSO mobility surveys for a given set of features.
}
\examples{
# We can use the nhanes dataset as an example even if it only contains population data
library(dplyr)
library(purrr)
mzmv_mean_map(
data = nhanes,
variable = c("annual_household_income", "annual_family_income"),
group_vars = c("gender", "interview_lang"),
weight = "weights"
)

}
