% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_ogd.R
\name{se_combn}
\alias{se_combn}
\title{Generate All Combinations of Strings from a Character Vector}
\usage{
se_combn(vars)
}
\arguments{
\item{vars}{A character vector of variable names.}
}
\value{
A list of character vectors, each representing a unique combination of the input strings.
The list includes the empty combination of length 0.
}
\description{
\code{se_combn()} is a helper function used internally to create all possible combinations
of a set of variables. It is typically used for generating grouped summary tables as per Open Government Data formats.
}
\examples{
vars <- letters[1:3]
se_combn(vars)

}
\keyword{internal}
