% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_dummy.R
\name{se_dummy}
\alias{se_dummy}
\title{Create Joint Dummy Variables from Multiple Categorical Columns}
\usage{
se_dummy(data, ..., sep = "_")
}
\arguments{
\item{data}{A data frame or tibble.}

\item{...}{One or more categorical columns (unquoted or quoted) whose combinations will be
used to generate joint dummy variables.}

\item{sep}{A character string to separate combined category names (default is "_").}
}
\value{
A tibble including all original data columns and additional dummy columns
representing all unique combinations of the provided categorical variables.
}
\description{
\code{se_dummy} is an internal helper function used to generate dummy (0/1) variables
based on the combinations of multiple categorical variables. It is primarily used by
\code{se_prop()} to support grouped dummy encoding.
}
\examples{
df <- tibble::tibble(gender = c("male", "female"), country = c("US", "Other"))
se_dummy(df, gender, country)

}
\keyword{internal}
