% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_dummy.R
\name{se_dummy}
\alias{se_dummy}
\title{Create Dummy Variables from a Categorical Variable}
\usage{
se_dummy(data, column)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{column}{Unquoted categorical column name to dummify.}
}
\value{
A tibble with the original data and newly created dummy variables added.
}
\description{
Create Dummy Variables from a Categorical Variable
}
\examples{
data <- data.frame(
  id = 1:5,
  category = c("A", "B", "A", "C", "B")
)
se_dummy(data, category)

}
