% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_mean.R
\name{se_mean_num}
\alias{se_mean_num}
\title{Estimate means of numeric variables in structural survey}
\usage{
se_mean_num(
  data,
  variable,
  condition = NULL,
  strata = "zone",
  weight,
  alpha = 0.05
)
}
\arguments{
\item{data}{Tibble}

\item{variable}{Character string, name of the feature whose mean we would like
to estimate}

\item{condition}{condition Vector of character strings, names of additional
stratification variables}

\item{strata}{Character string, name of the column containing the
strata/zones}

\item{weight}{Character string, name of the column containing the
weights}

\item{alpha}{Double, significance level. Default 0.05 for 95\% confidence interval.}
}
\value{
Tibble, with the following columns:
 \itemize{
 \item \code{occ}: true frequency in survey sample
 \item \code{average}: estimated mean
 \item \code{vhat}: estimated variance
 \item \code{stand_dev}: standard deviation
 \item \code{ci}: absolute confidence interval
 }
}
\description{
\code{se_mean_num()} estimates the averages of numeric variables along with the variance and confidence
intervals of FSO's structural survey (Strukturerhebung / relevé structurel).
}
\examples{
se_mean_num(
  data = nhanes,
  variable = "age",
  strata = "strata",
  weight = "weights",
  condition = "gender"
)

}
