% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_summarise.R
\name{se_summarise}
\alias{se_summarise}
\title{Summarise Structural Survey Data by Group}
\usage{
se_summarise(data, ..., weight, mh_col = "mh", Nh_col = "Nh")
}
\arguments{
\item{data}{A tibble or a data frame.}

\item{...}{One or more grouping variables. Can be passed unquoted (e.g., \code{gender}, \code{birth_country}) or programmatically using \code{!!!syms(c("gender", "birth_country"))}.}

\item{weight}{Unquoted or quoted name of the sampling weights column. For programmatic use
with a string variable (e.g., \code{wt <- "weights"}), use \code{!!sym(wt)} in the function call.}

\item{mh_col}{Character string, column name of number of observations (respondents) per group. Default is "mh".}

\item{Nh_col}{Character string, column name of estimated population per group. Default is "Nh".}
}
\value{
A tibble: the input \code{data} augmented with new columns \code{mh_col} and \code{Nh_col} (or their specified names).
}
\description{
\code{se_summarise()} calculates group-wise survey statistics, augmenting the input data with columns for the number of observations and estimated population per group.
}
\examples{
# One grouping variable
suppressPackageStartupMessages(library(dplyr))
se_summarise(nhanes, weight = weights, strata) |>
  glimpse()
# Two grouping variables
se_summarise(nhanes,
  weight = weights,
  mh_col = "mhc",
  Nh_col = "Nhc",
  strata, gender
)

}
