% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_summarise.R
\name{se_summarise}
\alias{se_summarise}
\title{Summarise Structural Survey Data by Group}
\usage{
se_summarise(data, ..., weight, mh_col = "mh", Nh_col = "Nh")
}
\arguments{
\item{data}{A data frame or tibble.}

\item{...}{One or more grouping variables. Can be passed unquoted (e.g., \code{gender}, \code{birth_country}) or programmatically using \code{!!!syms(c("gender", "birth_country"))}.}

\item{weight}{The name of the weight column. Can be passed unquoted, quoted, or programmatically using \code{!!sym(wt)} where \code{wt} is a character variable.}

\item{mh_col}{A character string specifying the name of the column for the number of observations (respondents) per group. Default is `"mh"`.}

\item{Nh_col}{A character string specifying the name of the column for the estimated population size per group. Default is `"Nh"`.}
}
\value{
A tibble containing the original data with two additional columns \code{mh_col} and \code{Nh_col} (or their specified names).
}
\description{
\code{se_summarise()} is a helper function for \code{\link[=se_total]{se_total()}}. It calculates survey-weighted group-level summaries by adding columns for the number
of observations and the estimated population size per group.
}
\examples{
# One grouping variable
se_summarise(nhanes, weight = weights, strata)
# Two grouping variables
se_summarise(nhanes,
  weight = weights,
  mh_col = "mhc",
  Nh_col = "Nhc",
  strata, gender
)

}
\keyword{internal}
