% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_total.R
\name{se_total}
\alias{se_total}
\title{Estimate Totals of Structural Survey}
\usage{
se_total(data, ..., weight, strata, alpha = 0.05)
}
\arguments{
\item{data}{A tibble or data frame.}

\item{...}{Additional unquoted grouping variables (e.g., \code{gender}, \code{marital_status}).}

\item{weight}{Unquoted column name of the column containing the weights.}

\item{strata}{Unquoted column name of the column containing the strata. Default is \code{zone}.}

\item{alpha}{Double, significance level. Default 0.05 for 95\% confidence interval.}
}
\value{
A tibble with estimates for all grouping column combinations, including:
\describe{
   \item{total}{population estimate.}
   \item{vhat}{estimated variance.}
   \item{occ}{true frequency in survey sample.}
   \item{stand_dev}{standard deviation.}
   \item{ci}{absolute confidence interval.}
   \item{ci_per}{percent confidence interval.}
   \item{ci_l}{Lower confidence interval bound.}
   \item{ci_u}{Upper confidence interval bound.}
 }
}
\description{
\code{se_total()} estimates the totals, variance, and confidence
intervals of FSO structural surveys.
}
\examples{
# One grouping variable
se_total(
  data = nhanes,
  weight = weights,
  strata = strata,
  gender
)
# Multiple grouping variables
se_total(
  data = nhanes,
  weight = weights,
  strata = strata,
  gender, marital_status, birth_country
)

}
\seealso{
\code{\link[=se_total_map]{se_total_map()}}
}
