% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_total_prop.R
\name{se_total_prop}
\alias{se_total_prop}
\title{Create a Table with Total and Proportion Estimates for Categorical Variables in Structural Survey}
\usage{
se_total_prop(data, ..., strata, weight, alpha = 0.05)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{...}{Optional grouping variables (unquoted).}

\item{strata}{The name of the strata variable (default is "zone").}

\item{weight}{The name of the weight variable.}

\item{alpha}{Significance level for confidence intervals. Default is 0.05.}
}
\value{
A tibble with joined total and proportion estimates.
}
\description{
\code{se_total_prop} is a wrapper function for \code{se_total()} and \code{se_prop()} which estimates totals and proportions for categorical variables.
}
\examples{
se_total_prop(
  data = nhanes,
  interview_lang,
  gender,
  birth_country,
  strata = strata,
  weight = weights
)

}
\seealso{
\code{\link[=se_total]{se_total()}}, \code{\link[=se_prop]{se_prop()}}.
}
